% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_polygon_sample_bias.R
\name{apply_polygon_sample_bias}
\alias{apply_polygon_sample_bias}
\title{Generate a sampling bias via a polygon}
\usage{
apply_polygon_sample_bias(observations, bias_area, bias_strength = 1)
}
\arguments{
\item{observations}{An sf object with POINT geometry.}

\item{bias_area}{An sf object with POLYGON geometry. The area in which the
sampling will be biased.}

\item{bias_strength}{A positive numeric value. The strength of the bias to
be applied in the biased area (as a multiplier). Above 1, area will be
oversampled. Below 1, area will be undersampled. For example, a value of 50
will result in 50 times more samples within the bias_area than outside.
Conversely, a value of 0.5 will result in half less samples within the
bias_area than outside.}
}
\value{
An sf object with POINT geometry with a bias_weight column
containing the sampling probability based on sampling bias.
}
\description{
This function adds a sampling bias weight column containing the sample
probability based on sampling bias within a polygon.
}
\examples{
library(sf)

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)

observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
  long = runif(n_points, xlim[1], xlim[2])
) \%>\%
  st_as_sf(coords = c("long", "lat"), crs = 3035)

# Create bias_area polygon overlapping two of the points
selected_observations <- st_union(observations_sf[2:3,])
bias_area <- st_convex_hull(selected_observations) \%>\%
  st_buffer(dist = 100) \%>\%
  st_as_sf()

bias_strength <- 2

apply_polygon_sample_bias(observations_sf, bias_area, bias_strength)

}
