% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_polygon_sampling_bias.R
\name{apply_polygon_sampling_bias}
\alias{apply_polygon_sampling_bias}
\title{Generate a sampling bias via a polygon}
\usage{
apply_polygon_sampling_bias(occurrences_sf, bias_area, bias_strength = 1)
}
\arguments{
\item{occurrences_sf}{An sf object with POINT geometry.}

\item{bias_area}{An sf object with POLYGON geometry. The area in which the
sampling will be biased.}

\item{bias_strength}{A positive numeric value. The strength of the bias to
be applied in the biased area (as a multiplier). Above 1, area will be
oversampled. Below 1, area will be undersampled. For example, a value of 50
will result in 50 times more samples within the bias_area than outside.
Conversely, a value of 0.5 will result in half less samples within the
bias_area than outside.}
}
\value{
An sf object with POINT geometry with a bias_weight column
containing the sampling probability based on sampling bias.
}
\description{
The function adds a sampling bias weight column containing the sample
probability based on bias strength within a given polygon.
}
\examples{
# Load packages
library(sf)
library(dplyr)
library(ggplot2)

# Set seed for reproducibility
set.seed(123)

# Simulate some occurrence data with coordinates and time points
num_points <- 10
occurrences <- data.frame(
  lon = runif(num_points, min = -180, max = 180),
  lat = runif(num_points, min = -90, max = 90),
  time_point = 0
  )

# Convert the occurrence data to an sf object
occurrences_sf <- st_as_sf(occurrences, coords = c("lon", "lat"))

# Create bias_area polygon overlapping at least two of the points
selected_observations <- st_union(occurrences_sf[2:3,])
bias_area <- st_convex_hull(selected_observations) \%>\%
  st_buffer(dist = 100) \%>\%
  st_as_sf()

occurrence_bias_sf <- apply_polygon_sampling_bias(
  occurrences_sf,
  bias_area,
  bias_strength = 2)
occurrence_bias_sf

# Visualise where the bias is
occurrence_bias_sf \%>\%
  mutate(bias_weight_f = as.factor(round(bias_weight, 3))) \%>\%
  ggplot() +
    geom_sf(data = bias_area) +
    geom_sf(aes(colour = bias_weight_f)) +
    ggtitle("Sampling Bias via Polygon")
}
\seealso{
Other detection: 
\code{\link{apply_manual_sampling_bias}()}
}
\concept{detection}
