% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_occurences.R
\name{custom_occurences}
\alias{custom_occurences}
\title{Function to generate temporally autocorrelated occurrences}
\usage{
custom_occurences(
  initial_average_abundance = NA,
  n_time_points = NA,
  population_change = NA,
  n_samp = NA
)
}
\arguments{
\item{initial_average_abundance}{A positive integer value indicating the
average number of occurrences to be simulated within the extend of \code{polygon}
at time point 1. This value will be used as mean of a Poisson distribution
(lambda parameter).}

\item{n_time_points}{A positive integer value indicating the number of time
points to simulate.}

\item{population_change}{A positive or negative numerical value fo the population growth rate}

\item{n_samp}{number of samples to take}
}
\value{
a matrix of number of occurrences with columns as time steps and rows as samples
}
\description{
Function to generate temporally autocorrelated occurrences
}
\examples{

library(dplyr)
library(ggplot2)
>>>>>>> 796d996c3d554dc4896fc41f9b64f745523b62e1
occ <- custom_occurences(initial_average_abundance = 20, n_time_points = 10,
                         population_change = 0.2, n_samp = 100)

occ \%>\%
  data.frame() \%>\%
  mutate(samp_id = 1:nrow(.) \%>\% as.factor()) \%>\%
  pivot_longer(-samp_id, names_to = "time", values_to = "n_occurrence") \%>\%
  mutate(time = str_remove(time, pattern = "X") \%>\% as.numeric()) \%>\%
  ggplot(aes(x = time, y = n_occurrence,col = samp_id)) +
  geom_line() +
  theme(legend.position = "none")
}
