% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_binormal_circle.R
\name{sample_from_binormal_circle}
\alias{sample_from_binormal_circle}
\title{Sample from a circle using the bivariate Normal distribution}
\usage{
sample_from_binormal_circle(observations, p_norm = 0.95, seed = NA)
}
\arguments{
\item{observations}{An sf object with POINT geometry and a
\code{coordinateUncertaintyInMeters} column. If this column is not present, the
function will assume no (zero meters) uncertainty around the observation
points.}

\item{p_norm}{A numeric value between 0 and 1. The proportion of all
possible samples from a a bivariate Normal distribution that fall within the
uncertainty circle. If no value is given, the default \code{p_norm} value is 0.95.}

\item{seed}{A positive numeric value. The seed for random number generation
to make results reproducible. If \code{NA} (the default), no seed is used.}
}
\value{
An sf object with POINT geometry containing the locations of the
sampled occurrences and a \code{coordinateUncertaintyInMeters} column containing
the coordinate uncertainty for each observation.
}
\description{
The function samples occurrences of a species within the uncertainty circle
around each observation assuming a bivariate Normal distribution with means
equal to the observation point and the variance equal to
(-\code{coordinateUncertaintyInMeters}^2) / (2 * log(1 - \code{p_norm})) such that
\code{p_norm} \% of all possible samples from this Normal distribution fall within
the uncertainty circle.
}
\examples{

library(simcuber)
library(sf)
library(dplyr)

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
coordinate_uncertainty <- rgamma(n_points, shape = 5, rate = 0.1)

observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
  long = runif(n_points, xlim[1], xlim[2]),
  coordinateUncertaintyInMeters = coordinate_uncertainty
) \%>\%
  st_as_sf(coords = c("long", "lat"), crs = 3035)

# Sample points within uncertainty circles according to normal rules
sample_from_binormal_circle(
  observations = observations_sf,
  p_norm = 0.95,
  seed = 123
)
}
