% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_occurrences.R
\name{sample_occurrences}
\alias{sample_occurrences}
\title{Sample occurrences from spatial random field}
\usage{
sample_occurrences(rs, ts)
}
\arguments{
\item{rs}{a raster object (terra)}

\item{ts}{vector with the number of occurrences by time step}
}
\value{
An sf object with POINT geometry
}
\description{
Draws occurrences (points) from a spatial random field (raster)
}
\examples{
library(spatstat)
library(sf)

r <- rast(ncol = 50, nrow = 50, xmin = 0, xmax = 50, ymin = 0, ymax = 50)
values(r) <- 1:ncell(r)
timeseries <- c(20, 40, 60)

pts_occ <- sample_occurrences(rs = r, ts = timeseries)

plot(r)
plot(pts_occ, add = TRUE, color = "black")

}
