% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_occurrences_from_raster.R
\name{sample_occurrences_from_raster}
\alias{sample_occurrences_from_raster}
\title{Sample occurrences from spatial random field}
\usage{
sample_occurrences_from_raster(rs, ts)
}
\arguments{
\item{rs}{a raster object (terra)}

\item{ts}{vector with the number of occurrences by time step}
}
\value{
An sf object with POINT geometry
}
\description{
Draws occurrences (points) from a spatial random field (raster)
}
\examples{
# Load packages
library(sf)
library(ggplot2)

# Create polygon
plgn <- st_polygon(list(cbind(c(5,10,8,2,3,5), c(2,1,7,9,5,2))))
ggplot() +
  geom_sf(data = plgn) +
  theme_minimal()

## Medium scale clustering
# Create the random field
rs_pattern_clustered <- create_spatial_pattern(
  polygon = plgn,
  resolution = 0.1,
  spatial_pattern = "clustered",
  seed = 123)

# Sample from random field
set.seed(123)
pts_occ_clustered <- sample_occurrences_from_raster(
  rs = rs_pattern_clustered,
  ts = rpois(1, 100))

ggplot() +
  tidyterra::geom_spatraster(data = rs_pattern_clustered) +
  geom_sf(data = pts_occ_clustered) +
  scale_fill_continuous(type = "viridis") +
  theme_minimal()

## Large scale clustering
# Create the random field
rs_pattern_large <- create_spatial_pattern(
  polygon = plgn,
  resolution = 0.1,
  spatial_pattern = 100,
  seed = 123)

# Sample from random field
set.seed(123)
pts_occ_large <- sample_occurrences_from_raster(
  rs = rs_pattern_large,
  ts = rpois(1, 100))

ggplot() +
  tidyterra::geom_spatraster(data = rs_pattern_large) +
  geom_sf(data = pts_occ_large) +
  scale_fill_continuous(type = "viridis") +
  theme_minimal()
}
\seealso{
Other occurrence: 
\code{\link{create_spatial_pattern}()},
\code{\link{simulate_random_walk}()},
\code{\link{simulate_timeseries}()}
}
\concept{occurrence}
