% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_random_walk.R
\name{simulate_random_walk}
\alias{simulate_random_walk}
\title{Simulate a random walk over time}
\usage{
simulate_random_walk(
  initial_average_occurrences = 50,
  n_time_points = 10,
  sd_step = 0.05,
  seed = NA
)
}
\arguments{
\item{initial_average_occurrences}{A positive integer value indicating the
average number of occurrences to be simulated within the extend of \code{polygon}
at time point 1. This value will be used as mean of a Poisson distribution
(lambda parameter).}

\item{n_time_points}{A positive integer value indicating the number of time
points to simulate.}

\item{sd_step}{A positive numeric value indicating the standard deviation of
the random steps.}

\item{seed}{A positive numeric value. The seed for random number generation
to make results reproducible. If \code{NA} (the default), no seed is used.}
}
\value{
A vector of integers of length n_time_points with the avarage number
of occurrences.
}
\description{
The function simulates occurrences of a species in a temporal extent.
}
\examples{

simulate_random_walk(
  initial_average_occurrences = 50,
  n_time_points = 10,
  sd_step = 1,
  seed = 123
)
}
\seealso{
Other occurrence: 
\code{\link{create_spatial_pattern}()},
\code{\link{sample_occurrences_from_raster}()},
\code{\link{simulate_timeseries}()}
}
\concept{occurrence}
