% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinate_uncertainty.R
\name{add_coordinate_uncertainty}
\alias{add_coordinate_uncertainty}
\title{Add coordinate uncertainty to observations}
\usage{
add_coordinate_uncertainty(observations, coords_uncertainty_meters = 25)
}
\arguments{
\item{observations}{An sf object with POINT geometry.}

\item{coords_uncertainty_meters}{A numeric value or a vector of numeric
values indicating the coordinate uncertainty associated with observations.}
}
\value{
An sf object with POINT geometry with an additional column
\code{coordinateUncertaintyInMeters}.
}
\description{
Adds a column to the observations sf object with the coordinate uncertainty
in meters.
}
\examples{

library(sf)
library(dplyr)

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
    long = runif(n_points, xlim[1], xlim[2])) \%>\%
    st_as_sf(coords = c("long", "lat"), crs = 3035)

 # provide a fixed uncertainty for all points
 add_coordinate_uncertainty(
   observations_sf,
   coords_uncertainty_meters = 1000
   )

# add variability in uncertainty. For example, using gamma distribution
add_coordinate_uncertainty(
  observations_sf,
  coords_uncertainty_meters = rgamma(n_points, shape = 5, rate = 0.1)
)
}
\seealso{
Other main: 
\code{\link{grid_designation}()},
\code{\link{sample_observations}()},
\code{\link{simulate_occurrences}()}
}
\concept{main}
