% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_pattern.R
\name{create_spatial_pattern}
\alias{create_spatial_pattern}
\title{Create spatial pattern}
\usage{
create_spatial_pattern(
  polygon,
  resolution,
  spatial_pattern = c("random", "clustered"),
  seed = NA,
  n_sim = 1
)
}
\arguments{
\item{polygon}{an sf object of geometry type POLYGON}

\item{resolution}{a numeric value defining the resolution of the raster cell}

\item{spatial_pattern}{define the spatial pattern. It could be a character string
"random" or "clustered", in which "random" is the default. The user is able to
provide a numeric value between 1 and 100. 1 is "random" and 5 is "clustered.
As large the number more broad is the size of the clusters area. See details.}

\item{seed}{integer. set a seed to randomization}

\item{n_sim}{number of simulations. Each simulation is a different layer in the raster.}
}
\value{
an object of class SpatRaster
}
\description{
It creates a raster with a spatial pattern for the area of a polygon.
}
\examples{

plgn <- st_polygon(list(cbind(c(5,10,8,2,3,5), c(2,1,7,9,5,2))))
plot(plgn)
# random pattern by default
spat_random <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  seed = 123,
  n_sim = 4)

plot(spat_random)

# built in clustered pattern

spat_clust<- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = "clustered",
  seed = 123,
  n_sim = 4)

plot(spat_clust)

# user defined spatial pattern
## small scale cluster

spat_small <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = 5,
  seed = 123,
  n_sim = 4)

plot(spat_small)

## medium scale cluster (the built in clustered pattern)
spat_medium <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = 10,
  seed = 123,
  n_sim = 4)

plot(spat_medium)

## large scale cluster
spat_large <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = 100,
  seed = 123,
  n_sim = 4)

plot(spat_large)


}
