% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_pattern.R
\name{create_spatial_pattern}
\alias{create_spatial_pattern}
\title{Create spatial pattern}
\usage{
create_spatial_pattern(
  polygon,
  resolution,
  spatial_pattern = c("random", "clustered"),
  seed = NA,
  n_sim = 1
)
}
\arguments{
\item{polygon}{An sf object of geometry type POLYGON}

\item{resolution}{A numeric value defining the resolution of the raster cell}

\item{spatial_pattern}{Define the spatial pattern. It could be a character
string "random" or "clustered", in which "random" is the default. The user
is able to provide a numeric value >= 1, where 1 is "random" and 10
is "clustered" built in options. As large the number more broad is the size
of the clusters area. See details.}

\item{seed}{The seed for random number generation to make results
reproducible. If \code{NA} (the default), no seed is used.}

\item{n_sim}{Number of simulations. Each simulation is a different layer in
the raster. Default 1.}
}
\value{
An object of class SpatRaster with a spatial pattern for the area of
the given polygon.
}
\description{
It creates a raster with a spatial pattern for the area of a polygon.
}
\details{
the \code{spatial_pattern} argument change the range parameter of the
spherical variogram model. \code{spatial_pattern = 1} means the range has
the same size of the grid cell, which is defined in \code{resolution}
argument. We use the functon \code{\link[gstat:vgm]{gstat::vgm()}} to implement the
spherical variogram model
}
\examples{

library(sf)
library(terra)

plgn <- st_polygon(list(cbind(c(5,10,8,2,3,5), c(2,1,7,9,5,2))))
plot(plgn)
# random pattern by default
spat_random <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  seed = 123)

plot(spat_random, main = "random pattern")

# built in clustered pattern

spat_clust<- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = "clustered",
  seed = 123)

plot(spat_clust, main = "built in clustered pattern")

# user defined spatial pattern
## small scale cluster

spat_small <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = 5,
  seed = 123)

plot(spat_small, main = "small scale clustered pattern")

## medium scale cluster (the built in clustered pattern)
spat_medium <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = 10,
  seed = 123)

plot(spat_medium, main = "medium scale clustered pattern")

## large scale cluster
spat_large <- create_spatial_pattern(
  plgn,
  resolution = 0.1,
  spatial_pattern = 100,
  seed = 123)

plot(spat_large, main = "large scale clustered pattern")

}
\seealso{
\code{\link[gstat:vgm]{gstat::vgm()}} and its \code{range} argument
}
