% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_occurrences.R
\name{simulate_occurrences}
\alias{simulate_occurrences}
\title{Simulate occurrences within a spatiotemporal scope}
\usage{
simulate_occurrences(
  plgn,
  initial_average_abundance = 50,
  spatial_autocorr = c("random", "clustered"),
  n_time_points = 1,
  temporal_function = NA,
  spatiotemporal_autocorr = NA,
  seed = NA
)
}
\arguments{
\item{plgn}{An sf object with POLYGON geometry indicating the spatial
extend to simulate occurrences.}

\item{initial_average_abundance}{A positive integer value indicating the
average number of occurrences to be simulated within the extend of \code{polygon}
at time point 1. This value will be used as mean of a Poisson distribution
(lambda parameter).}

\item{spatial_autocorr}{Define the spatial pattern. It could be a character
string \code{"random"} or \code{"clustered"}, in which \code{"random"} is the default.
The user is able to provide a numeric value >= 1 (1 is "random" and
10 is "clustered"). A larger number means a broader size of the clusters
area. See details.}

\item{n_time_points}{A positive integer value indicating the number of time
points to simulate.}

\item{temporal_function}{\code{NA} (default), or a function which generates
a trend in abundance over time. Only used if \code{n_time_points > 1}. By default,
the function will sample \code{n_time_points} times from a Poisson
distribution with average (lambda) \code{initial_average_occurrences}. When a
function is specified (e.g. the internal \code{simulate_random_walk()} function).}

\item{spatiotemporal_autocorr}{A numeric value between indicating the
strength of spatiotemporal autocorrelation.}

\item{seed}{A positive numeric value. The seed for random number generation
to make results reproducible. If \code{NA} (the default), no seed is used.}
}
\value{
An sf object with POINT geometry containing the locations of the
simulated occurrences and a \code{time_point} column containing the time point
associated with each occurrence.
}
\description{
The function simulates occurrences of a species within a given spatial
and/or temporal extend.
}
\examples{
# Load packages
library(sf)
library(ggplot2)

# Create polygon
plgn <- st_polygon(list(cbind(c(5,10,8,2,3,5), c(2,1,7,9,5,2))))
ggplot() +
  geom_sf(data = plgn) +
  theme_minimal()

## Random spatial pattern with 4 time points
occ_sf <- simulate_occurrences(
  plgn,
  n_time_points = 4,
  initial_average_abundance = 100,
  seed = 123)

ggplot() +
 geom_sf(data = occ_sf) +
 geom_sf(data = plgn, fill = NA) +
 facet_wrap("time_point") +
 labs(
      title = "Occurrences with random\nspatial and temporal pattern",
      subtitle = "4 time steps") +
 theme_bw()

## Clustered spatial pattern with 4 time points
occ_sf_100 <- simulate_occurrences(
  plgn,
  spatial_autocorr = 100,
  n_time_points = 4,
  initial_average_abundance = 100,
  seed = 123)

ggplot() +
  geom_sf(data = occ_sf_100) +
  geom_sf(data = plgn, fill = NA) +
  facet_wrap("time_point") +
  labs(
       title = "Occurrences with structured\nspatial and temporal pattern",
       subtitle = "4 time steps") +
  theme_bw()
}
\seealso{
Other main: 
\code{\link{add_coordinate_uncertainty}()},
\code{\link{grid_designation}()},
\code{\link{sample_observations}()}
}
\concept{main}
